renaming etherate_mt.* to main.* as part of issue #39
